Progres in the way Finally something is working + repaired move class to be fully functional
ToDo:
- numeric menu :D
- import path from settings
- tab autocomplete (meaby replace command read line with someone else code?)
- shortcuts for --commands (or tab to be faster)

+++ Move.cs
@@ -26,7 +26,7 @@ public class MoveFilesClass
         string whatToMove = "";
         string extension = "";
 
-        void commandChecker(string command, string[] commands, int index)
+        bool commandChecker(string command, string[] commands, int index)
         {
             if (command.Contains("--"))
             {
@@ -34,12 +34,17 @@ public class MoveFilesClass
                     create = command;
                 else if (command == "--subfolders")
                     subFolders = command;
-                else if (command == whatToMove)
+                else if (command == "--whatmove")
                 {
                     whatToMove = commands[index + 1];
                 }
                 else if (command == "--extension")
                     extension = commands[index + 1];
+                else if (command.Contains("--"))
+                {
+                    Console.WriteLine($"\nCommand {command} is not recognized");
+                    return false;
+                }
             }
 
             if (commands.Length >= 3)
@@ -50,36 +55,31 @@ public class MoveFilesClass
 
             if (commands.Length >= 3)
             {
-                commands[2] = @"C:\Users\Nowe Jakubki\Pictures\docelowe";
+
+                commands[2] = @"C:\Users\Nowe Jakubki\Pictures\docelowe\czy stworzy";
                 finalFilePath = commands[2];
             }
+            return true;
 
         }
 
+        
+
         for (int i = 1; i < commands.Length; i++)
         {
-            commandChecker(commands[i], commands, i);
+            if (commandChecker(commands[i], commands, i) == false)
+                return;
         }
 
 
         string result = MoveFiles(filePath, finalFilePath, create, subFolders, whatToMove, extension);
-        Console.WriteLine($"\n {result}"); 
+        Console.WriteLine($"\n {result}");
 
     }
 
+
     public string MoveFiles(string filePath, string finalFilePath, string create = "", string subFolders = "", string whatToMove = "", string extension = "", params string[] fileName)
     {
-        //filename -> ilość plików
-        //subfolders
-        //extension -> declares if it's file or not 
-        //what to move - file or directory
-
-
-
-        //na start sprawdzić
-        // - czy istnieją obie ścieżki + czy istnieje create
-        // - czy źródłowa ścieżka ma pliki
-
         //
         //source path checker
         if (Directory.Exists(filePath) == false)
@@ -91,62 +91,68 @@ public class MoveFilesClass
         //final file path checker / creator
         if (finalFilePath == "")
         {
-            if(Create(create) == false)
+            Create(create);
+            return "Moving operation didn't end succesfully\n\n";
+        }
+        else
+        {
+            if (Create(create) == false)
             {
-                return "Final path does not exist and --create option was not writtend";
+                return "Moving operation didn't end succesfully\n\n";
             }
         }
-        else if (finalFilePath == "")
-        {
-            return "You didn't write final path ";
-        }
-        
+
+
+
         string[] files = Search_and_SubFolders(subFolders);
         WhatFileToMove(files, whatToMove, extension);
-        return "Files moved";
+        return "\nMoving operation has done\n\n";
 
         bool Create(string create)
         {
             if (Directory.Exists(finalFilePath) == false)
             {
-                if (create != "")
+                if (create != "" && finalFilePath != "")
                 {
-                    Directory.CreateDirectory(finalFilePath);
-                    return true;
+                    try
+                    {
+                        Directory.CreateDirectory(finalFilePath);
+                        return true;
+
+                    }
+                    catch (Exception)
+                    {
+                        Console.WriteLine("\nCreating directory has failed");
+                        throw;
+                    }
 
                 }
                 else
-                    Console.WriteLine("target path does not exist and you didn't wanted to create one");
+                    Console.WriteLine("\ntarget path does not exist, you didn't wanted to create one or you didn't put final path");
                 return false;
             }
             else
             {
-                Console.WriteLine("Directory " + finalFilePath + " already exist");
+
+                if (create == "")
+                {
+                    return true;
+                }
+
+                Console.WriteLine();
+                Console.Write("\n\nDirectory " + finalFilePath + " already exist ->");
                 int n = 0;
                 do
                 {
                     n++;
 
-                } while (Directory.Exists(@$"{finalFilePath}\({n})") == false);
+                } while (Directory.Exists(@$"{finalFilePath} ({n})"));
 
-                Directory.CreateDirectory(@$"{finalFilePath}\({n})");
+                Directory.CreateDirectory(@$"{finalFilePath} ({n})");
+                finalFilePath = @$"{finalFilePath} ({n})";
+                Console.Write($" we created folder \"{Path.GetFileName(finalFilePath)}\"\n");
                 return true;
 
-            }
-
-        }
-        string[] Search_and_SubFolders(string subFolders)
-        {
-
-            List<string[]> tempSubFolders = new List<string[]>();
-
-            if (subFolders == "")
-            {
-                return Directory.GetFileSystemEntries(filePath, "*", SearchOption.TopDirectoryOnly);
-            }
-            else
-            {
-                return Directory.GetFileSystemEntries(filePath, "*", SearchOption.AllDirectories);
             }
 
         }
@@ -158,87 +164,41 @@ public class MoveFilesClass
             {
                 int counter = 0;
 
-                if (extension == "pdf")
-                {
-                    for (int i = 0; i < filesArr.Length; i++)
-                    {
-                        if (Path.GetExtension(filesArr[i]) == ".pdf")
-                        {
-                            string fileName = Path.GetFileName(filesArr[i]);
-                            File.Move(filesArr[i], @$"{finalFilePath}\{fileName}");
-                            counter++;
-                        }
-                    }
-                    if (counter == 0)
-                        Console.WriteLine("PDF files not found");
-                    else
-                        Console.WriteLine($"{counter} files were moved");
 
-                }
-                else if (extension == "jpg")
+                if (extension.Substring(0, 1) == ".")
                 {
-                    for (int i = 0; i < filesArr.Length; i++)
-                    {
-                        if (Path.GetExtension(filesArr[i]) == ".jpg")
-                        {
-                            string fileName = Path.GetFileName(filesArr[i]);
-                            File.Move(filesArr[i], @$"{finalFilePath}\{fileName}");
-                            counter++;
-                        }
-                    }
-                    if (counter == 0)
-                        Console.WriteLine("JPG files not found");
-                    else
-                        Console.WriteLine($"{counter} files were moved");
+                    extension = extension.Substring(1, extension.Length - 1);
+                    Console.WriteLine(extension);
                 }
-                else if (extension == "png")
+
+                for (int i = 0; i < filesArr.Length; i++)
                 {
-                    for (int i = 0; i < filesArr.Length; i++)
+                    if (Path.GetExtension(filesArr[i]) == $".{extension}")
                     {
-                        if (Path.GetExtension(filesArr[i]) == ".png")
+                        string fileName = Path.GetFileName(filesArr[i]);
+
+                        if (File.Exists($@"{finalFilePath}\{fileName}"))
                         {
-                            string fileName = Path.GetFileName(filesArr[i]);
-                            File.Move(filesArr[i], @$"{finalFilePath}\{fileName}");
-                            counter++;
+                            int j = 0;
+                            do
+                            {
+                                j++;
+                            } while (File.Exists($@"{finalFilePath}\{fileName} ({j})"));
+                            File.Move(filesArr[i], $@"{finalFilePath}\{fileName} ({j})");
                         }
-                    }
-                    if (counter == 0)
-                        Console.WriteLine("PNG files not found");
-                    else
-                        Console.WriteLine($"{counter} files were moved");
-                }
-                else if (extension == "xlsx" ^ extension == "excel")
-                {
-                    for (int i = 0; i < filesArr.Length; i++)
-                    {
-                        if (Path.GetExtension(filesArr[i]) == ".xlsx")
+                        else
                         {
-                            string fileName = Path.GetFileName(filesArr[i]);
-                            File.Move(filesArr[i], @$"{finalFilePath}\{fileName}");
-                            counter++;
+                            File.Move(filesArr[i], $@"{finalFilePath}\{fileName}");
                         }
+                        counter++;
                     }
-                    if (counter == 0)
-                        Console.WriteLine("Excel files not found");
-                    else
-                        Console.WriteLine($"{counter} files were moved");
                 }
-                else if (extension == "docx" ^ extension == "word")
-                {
-                    for (int i = 0; i < filesArr.Length; i++)
-                    {
-                        if (Path.GetExtension(filesArr[i]) == ".docx")
-                        {
-                            string fileName = Path.GetFileName(filesArr[i]);
-                            File.Move(filesArr[i], @$"{finalFilePath}\{fileName}");
-                            counter++;
-                        }
-                    }
-                    if (counter == 0)
-                        Console.WriteLine("Word files not found");
-                    else
-                        Console.WriteLine($"{counter} files were moved");
-                }
+
+                if (counter == 0)
+                    Console.WriteLine($"\nExtension '{extension} was not found'");
+                else
+                    Console.WriteLine($"\n\nMoved files: {counter}");
+
             }
 
             if (extension != "")
@@ -259,21 +219,51 @@ public class MoveFilesClass
                         string fileName = Path.GetFileName(filesArr[i]);
 
                         if (File.Exists(filesArr[i]))
-                            File.Move(filesArr[i], @$"{finalFilePath}\{fileName}");
+                        {
+                            if (File.Exists($@"{finalFilePath}\{fileName}"))
+                            {
+                                int j = 0;
+                                do
+                                {
+                                    j++;
+                                } while (File.Exists($@"{finalFilePath}\{fileName} ({j})"));
+                                File.Move(filesArr[i], $@"{finalFilePath}\{fileName} ({j})");
+                            }
+                            else
+                            {
+                                File.Move(filesArr[i], @$"{finalFilePath}\{fileName}");
+                            }
+                        }
+
                     }
                 }
                 else if (whatToMove == "directory" ^ whatToMove == "dir" ^ whatToMove == "folder" ^ whatToMove == "folders")
                 {
+                    filesArr = Directory.GetDirectories(filePath);
                     for (int i = 0; i < filesArr.Length; i++)
                     {
-
                         string fileName = Path.GetFileName(filesArr[i]);
 
                         if (Directory.Exists(filesArr[i]))
-                            Directory.Move(filesArr[i], @$"{finalFilePath}\{fileName}");
+                        {
+                            if (Directory.Exists($@"{finalFilePath}\{fileName}"))
+                            {
+                                int j = 0;
+                                do
+                                {
+                                    j++;
+                                } while (Directory.Exists($@"{finalFilePath}\{fileName} ({j})"));
+                                Directory.Move(filesArr[i], $@"{finalFilePath}\{fileName} ({j})");
+                            }
+                            else
+                            {
+                                Directory.Move(filesArr[i], @$"{finalFilePath}\{fileName}");
+                            }
 
+                        }
                     }
                 }
+
             }
             else
             {
@@ -281,26 +271,43 @@ public class MoveFilesClass
                 {
                     string fileName = Path.GetFileName(filesArr[i]);
 
-                    Directory.Move(filesArr[i], @$"{finalFilePath}\{fileName}");
+                    if (Directory.Exists($@"{finalFilePath}\{fileName}")  ^ File.Exists($@"{finalFilePath}\{fileName}"))
+                    {
+                        int j = 0;
+                        do
+                        {
+                            j++;
+                        } while (Directory.Exists($@"{finalFilePath}\{fileName} ({j})") ^ File.Exists($@"{finalFilePath}\{fileName} ({j})"));
+                        Directory.Move(filesArr[i], $@"{finalFilePath}\{fileName} ({j})");
+                    }
+                    else
+                    {
+                        Directory.Move(filesArr[i], @$"{finalFilePath}\{fileName}");
+                    }
                 }
+
             }
         }
 
-        //tu utworzyć osobne dla files -> wiadmom czego szukać to nie będzie potrzeba automatyzacji
+        string[] Search_and_SubFolders(string subFolders)
+        {
 
+            List<string[]> tempSubFolders = new List<string[]>();
 
-
-
-
-
-
-
-
-
-
+            //if (subFolders == "")
+            //{
+            return Directory.GetFileSystemEntries(filePath, "*", SearchOption.TopDirectoryOnly);
+            //}
+            // else
+            // {
+            //     return Directory.GetFileSystemEntries(filePath, "*", SearchOption.AllDirectories);
+            // }
+            //
+            // commented due to lack of need to move subfolders (they move with parent... )
+            // 
+        }
 
     }
-
 
 
 }
